Switch to new sqrt in NetEQ
Switched to WebRtcSpl_SqrtFloor instead of WebRtcSpl_Sqrt in
NetEQ. The output is not bit-exact, but subjective listening
tests show no audible difference. Analysis shows that almost
all of the difference is in changed delay.

The reference file for NetEQ's unit test was updated.

Review URL: http://webrtc-codereview.appspot.com/139019