Add Chromium's src/buildtools to DEPS.
GN for WebRTC was broken by the depot_tools change in
https://codereview.chromium.org/341533006/ that changes
the gn.py wrapper to use GN in src/buildtools instead of the
previous location in tools/gn/bin.

This buildtools repo was added for Chromium in
https://codereview.chromium.org/281863002 and the hooks were
updated in https://codereview.chromium.org/340153002

This adds the buildtools dir and updates our download hooks.

BUG=webrtc:3441
TEST=Locally running GN (trybots currently cannot handle DEPS changes properly)
R=niklas.enbom@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/15929004